# doc/api.yml
openapi: 3.0.3
info:
  title: HEI - Grade Management API
  description: >
    API for managing grades at the Higher Institute of Computer Science (HEI).
    Covers authentication (student/teacher/admin login), the academic repository
    (cohorts, groups, courses, EL/TN tracks), exams and grades (with history),
    grade transcripts (PDF via S3 + Poja email, asynchronous), and the graduate
    list (Excel export).
  version: "2.0.0"
servers:
  - url: https://ykc5vnpsxpkahbcvse6pnvowt40mupoe.lambda-url.eu-west-3.on.aws/
    description: Poja deployment environment (AWS Lambda)
...
security:
  - bearerAuth: []
paths:
...
  /students:
    post:
      tags: [Students]
      summary: Add a student
      description: >
        Restricted to ADMIN role. The email is NOT provided by the caller: it is
        generated server-side as hei.firstname@student.com. An initial password is
        generated as well (to be reset on first login).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StudentCreateRequest" }
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Student" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /students/{id}:
    delete:
      tags: [Students]
      summary: Delete a student
      description: Restricted to ADMIN role. Grade history is preserved.
      parameters:
        - $ref: "#/components/parameters/StudentId"
      responses:
        "204": { description: Student deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /students/{id}/group:
    patch:
      tags: [Students]
      summary: Change a student's group
      description: >
        Restricted to ADMIN role. Closes the current group history and opens a new entry.
        The student's grades are never affected by this change.
      parameters:
        - $ref: "#/components/parameters/StudentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GroupChangeRequest" }
      responses:
        "200":
          description: Group updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Student" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /students/{id}/parcours:
    patch:
      tags: [Students]
      summary: Set a student's track (EL/TN)
      description: Restricted to ADMIN role. Applicable from the 4th semester onwards.
      parameters:
        - $ref: "#/components/parameters/StudentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [track]
              properties:
                track: { $ref: "#/components/schemas/Track" }
      responses:
        "200":
          description: Track updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Student" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /students/{id}/grades:
    get:
      tags: [Grades]
      summary: View a student's grades
      description: STUDENT (their own), TEACHER (their courses only), ADMIN (any student).
      parameters:
        - $ref: "#/components/parameters/StudentId"
        - in: query
          name: academicYear
          schema: { type: integer }
      responses:
        "200":
          description: Grade details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudentGrades" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /students/{id}/grades/history:
    get:
      tags: [Grades]
      summary: History of grade modifications for a student
      description: ADMIN, or STUDENT for themselves (read-only).
      parameters:
        - $ref: "#/components/parameters/StudentId"
      responses:
        "200":
          description: Grade history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GradeHistoryEntry" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /students/{id}/years/{year}/summary:
    get:
      tags: [Transcripts]
      summary: Annual summary (grades, average, credits, status)
      description: >
        STUDENT (their own), TEACHER (their courses, partial view), ADMIN (any student).
        The returned status is PROVISIONAL or COMPLETE depending on whether all expected
        grades for the year have been entered.
      parameters:
        - $ref: "#/components/parameters/StudentId"
        - in: path
          name: year
          required: true
          schema: { type: integer, minimum: 1, maximum: 3 }
      responses:
        "200":
          description: Annual summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/YearSummary" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /students/{id}/transcript:
    post:
      tags: [Transcripts]
      summary: Generate and send grade transcript (PDF, asynchronous)
      description: >
        STUDENT (only for themselves) or ADMIN (any student). Asynchronously triggers:
        PDF generation (provisional or complete), upload to AWS S3, then email sending
        (Poja email service) containing the link. HTTP response is immediate (202);
        the document and email arrive subsequently.
      parameters:
        - $ref: "#/components/parameters/StudentId"
        - in: query
          name: academicYear
          schema: { type: integer }
        - in: query
          name: recipientEmail
          required: false
          schema: { type: string, format: email }
          description: Optional destination email. If omitted, the student's institutional email is used.
      responses:
        "202":
          description: Generation request accepted (asynchronous processing)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TranscriptRequestAck" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    StudentId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid request (e.g., business rule violation)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Access denied for the current role
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
...
    Track:
      type: string
      enum: [EL, TN]
      nullable: true
      description: Software Ecosystem (EL) or Digital Transformation (TN). Null before the 4th semester.
    CursusStatus:
      type: string
      enum: [ACTIVE, GRADUATED, DROPPED_OUT]
    ReportStatus:
      type: string
      enum: [PROVISIONAL, COMPLETE]
...
    GroupChangeRequest:
      type: object
      required: [newGroupId, effectiveDate]
      properties:
        newGroupId: { type: string, format: uuid }
        effectiveDate: { type: string, format: date }
    Student:
      type: object
      properties:
        id: { type: string, format: uuid }
        ref: { type: string, example: "STD00123" }
        name: { type: string }
        firstname: { type: string }
        email:
          type: string
          format: email
          description: Server-generated, read-only (hei.firstname@student.com).
          readOnly: true
        cohortId: { type: string, format: uuid }
        groupId: { type: string, format: uuid }
        track: { $ref: "#/components/schemas/Track" }
        status: { $ref: "#/components/schemas/CursusStatus" }
    StudentCreateRequest:
      type: object
      required: [name, firstname, cohortId]
      properties:
        name: { type: string }
        firstname: { type: string }
        address: { type: string }
        birthdate: { type: string, format: date }
        cohortId: { type: string, format: uuid }
        groupId:
          type: string
          format: uuid
          description: Optional — if absent, automatic assignment.
...
    GradeHistoryEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        gradeId: { type: string, format: uuid }
        oldValue: { type: number, format: double, nullable: true }
        newValue: { type: number, format: double }
        changedBy: { type: string, format: uuid }
        changedAt: { type: string, format: date-time }
        reason: { type: string }
    CourseGradeSummary:
      type: object
      properties:
        courseId: { type: string, format: uuid }
        courseRef: { type: string }
        courseTitle: { type: string }
        credit: { type: integer }
        finalGrade: { type: number, format: double }
    StudentGrades:
      type: object
      properties:
        studentId: { type: string, format: uuid }
        academicYear: { type: integer }
        courses:
          type: array
          items: { $ref: "#/components/schemas/CourseGradeSummary" }
    YearSummary:
      type: object
      properties:
        studentId: { type: string, format: uuid }
        year: { type: integer }
        courses:
          type: array
          items: { $ref: "#/components/schemas/CourseGradeSummary" }
        overallAverage: { type: number, format: double }
        totalCredits: { type: integer, example: 60 }
        status: { $ref: "#/components/schemas/ReportStatus" }
    TranscriptRequestAck:
      type: object
      properties:
        requestId: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        status:
          type: string
          enum: [ACCEPTED]
        message:
          type: string
          example: "Generation in progress, an email will be sent with the document link."
...